+++ packages/at_commons/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.0.10
 - fix: Deprecate apkam in PkamAuthMode enum
+- fix: Add a "force" variable to enroll_verb_builder to propagate enroll:revoke:force value
 ## 4.0.9
 - feat: enroll verb syntax change for enroll:revoke:force and added new exception AtEnrollmentRevokeException
 ## 4.0.8